+++ src/data-access/broker/broker.interface.ts
@@ -6,7 +6,7 @@
  * (Department of Information and Computing Sciences)
  */
 
-import { UseIsAuthorizedState } from "../features/authentication/AuthSlice";
+import { UseIsAuthorizedState } from "../authentication/AuthSlice";
 import { SendMessageI } from "./types";
 
 /**
