+++ src/data-access/broker/hooks.ts
@@ -8,8 +8,10 @@
 
 import { useState, useCallback, useEffect } from "react";
 import type { MpBackendAction, MpBackendMessage } from "./types";
-import { Addon, AddonCategory } from "../types/AddOnTypes";
-import { createBroker } from "../utils/mocking-utils";
+// import { Addon, AddonCategory } from "../../types/addon";
+import type { Addon, AddonCategory } from "/types/addon";
+// import { createBroker } from "../../test/mocking-utils";
+import { createBroker } from "/test/mocking-utils";
 
 const Broker = createBroker();
