+++ src/data-access/hooks.ts
@@ -17,10 +17,7 @@
 /* eslint-disable @typescript-eslint/no-restricted-imports */
 import { useDispatch, useSelector } from "react-redux";
 import type { AppDispatch, RootState } from "./store";
-import {
-  UseIsAuthorizedState,
-  authState
-} from "../features/authentication/AuthSlice";
+import { UseIsAuthorizedState, authState } from "./authentication/AuthSlice";
 
 // Use throughout your app instead of plain `useDispatch` and `useSelector`
 export const useAppDispatch = useDispatch.withTypes<AppDispatch>();
