+++ src/stories/RTKError.stories.tsx
@@ -8,7 +8,7 @@
 
 import type { Meta, StoryObj } from "@storybook/react";
 import { Provider } from "react-redux";
-import { store } from "../app/store";
+import { store } from "../data-access/store";
 import { MemoryRouter } from "react-router-dom";
 import RTKError from "../components/RTKError";
